requirements.txt
pyside==1.2.4
pyephem==3.7.6.0
pyserial==3.0.1